+++ zapisy/apps/offer/vote/templates/vote/summary.html
@@ -1,6 +1,7 @@
 {% extends "offer/base.html" %}
 
 {% load proposal_status %}
+{% load vote_filters %}
 {% load render_bundle from webpack_loader %}
 
 {% block main-subtitle %}Podsumowanie głosowania{% endblock %}
